Update EXISTING patch version for each minor ruby
Previously, we only looked in the toplevel rvm key.  We have a few
repositories that also have excludes with the same exact rvm ruby
version format that need to be updated.

Additionally, we allowed you to add new ruby minor versions to the
.travis.yml.  This feels dangerous.  I'd like this script to be simple
and just bump each specified minor version to the latest patch.

Finally, you can now specify ruby versions with components > 9.
Previously, it would yell at you if you tried to specify 2.5.11.

Here's an example showing that I missed it last time:
https://github.com/ManageIQ/manageiq-ui-classic/pull/5595

Here's one that was done using the blaster outer from this commit:
https://github.com/ManageIQ/manageiq/pull/18780